codeforces

+++ B_Two_Large_Bags.cpp
@@ -1,31 +1,31 @@
-#include<bits/stdc++.h>
+#include <bits/stdc++.h>
 using ll = long long;
 
-void solve() {
+void solve(){
     ll n;
-    std::cin>>n;
-    std::vector<int> a(n);
-    for(int i=0;i<n; i++) {
-        std::cin>>a[i];
+    std::cin >> n;
+    std::vector<ll> a(n);
+    for (int i = 0; i < n; i++){
+        std::cin >> a[i];
     }
     std::sort(a.begin(), a.end());
-    int mx = 0;
-    for (int i = 0; i < n; i += 2) {
-        if (std::max(mx, a[i]) != std::max(mx, a[i + 1])) {
-            std::cout<<"No"<<"\n";
+    ll mx = 0;
+    for (int i = 0; i < n; i += 2){
+        if (std::max(mx, a[i]) != std::max(mx, a[i + 1])){
+            std::cout << "No" << "\n";
             return;
         }
         mx = std::max(mx, a[i]) + 1;
     }
-    std::cout<<"Yes"<<"\n";
+    std::cout << "Yes" << "\n";
 }
 
-int main() {
+int main(){
     std::ios::sync_with_stdio(false);
     std::cin.tie(nullptr);
     int t;
-    std::cin>>t;
-    while (t--) {
+    std::cin >> t;
+    while (t--){
         solve();
     }
 }
